Functions to postInit
Functions will now be called in postInit and preInit. This means initPlayerLocal.sqf and initServer.sqf files no longer need to be present and copied over in the actual mission folders. Also allows me to initialize eventHandlers in preInit, making sure that the player has them when events are called on start, like JIP events.